refactor: rename `settings` module to `options`
