+++ init.lua
@@ -53,7 +53,7 @@ g.localleader = '\\'
 
 require('plugins') -- Plugin definitions
 require('colors') -- Pretty rainbows!!
-require('settings') -- General Neovim Settings
+require('options') -- General Neovim Settings
 require('filetypes') -- Filetype associations
 require('utils') -- General utility functions
 require('commands') -- Command abbreviations
